Finalização das funcionalidades do blog, inclusão de autenticação de usuário e camadas de proteção e acréscimo de upload de imagens.

+++ blog-php/app/Http/Controllers/PostsController.php
@@ -3,11 +3,22 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Storage;
 use App\Models\Post;
 //use DB;
 
 class PostsController extends Controller
 {
+    /**
+     * Create a new controller instance.
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        $this->middleware('auth', ['except' => ['index', 'show']]);
+    }
+
     /**
      * Display a listing of the resource.
      *
@@ -40,12 +51,30 @@ class PostsController extends Controller
     {
         $this->validate($request, [
             'title' => 'required',
-            'description' => 'required'
+            'description' => 'required',
+            'cover_image' => 'image|nullable|max:1999'
         ]);
+        //handle the file upload
+        if($request->hasFile('cover_image')) {
+            //get filename with extension
+            $fileNameWithExt = $request->file('cover_image')->getClientOriginalName();
+            //get just filename
+            $filename = pathinfo($fileNameWithExt, PATHINFO_FILENAME);
+            //get just ext
+            $extension = $request->file('cover_image')->getClientOriginalExtension();
+            //filename to store
+            $fileNameToStore = $filename.'_'.time().'.'.$extension;
+            //upload image
+            $path = $request->file('cover_image')->storeAs('public/cover_images', $fileNameToStore);
+        } else {
+            $fileNameToStore = 'noimage.jpg';
+        }
 
         $post = new Post;
         $post->title = $request->input('title');
         $post->description = $request->input('description');
+        $post->user_id = auth()->user()->id;
+        $post->cover_image = $fileNameToStore;
         $post->save();
 
         return redirect('/posts')->with('success', 'Post created successfully!');
@@ -72,6 +101,11 @@ class PostsController extends Controller
     public function edit($id)
     {
         $post = Post::find($id);
+        if(auth()->user()->id !== $post->user_id)
+        {
+            return redirect('/posts')->with('error', 'Unauthorized page.');
+        }
+
         return view('posts.edit')->with('post', $post);
     }
 
@@ -86,12 +120,30 @@ class PostsController extends Controller
     {
         $this->validate($request, [
             'title' => 'required',
-            'description' => 'required'
+            'description' => 'required',
+            'cover_image' => 'image|nullable|max:1999'
         ]);
+        //handle the file upload
+        if($request->hasFile('cover_image')) {
+            //get filename with extension
+            $fileNameWithExt = $request->file('cover_image')->getClientOriginalName();
+            //get just filename
+            $filename = pathinfo($fileNameWithExt, PATHINFO_FILENAME);
+            //get just ext
+            $extension = $request->file('cover_image')->getClientOriginalExtension();
+            //filename to store
+            $fileNameToStore = $filename.'_'.time().'.'.$extension;
+            //upload image
+            $path = $request->file('cover_image')->storeAs('public/cover_images', $fileNameToStore);
+        }
 
         $post = Post::find($id);
         $post->title = $request->input('title');
         $post->description = $request->input('description');
+        if($request->hasFile('cover_image'))
+        {
+            $post->cover_image = $fileNameToStore;
+        }
         $post->save();
 
         return redirect('/posts')->with('success', 'Post updated successfully!');
@@ -106,6 +158,14 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
+        if(auth()->user()->id !== $post->user_id)
+        {
+            return redirect('/posts')->with('error', 'Unauthorized page.');
+        }
+        if($post->cover_image != 'noimage.jpg')
+        {
+            Storage::delete('public/cover_images/'.$post->cover_image);
+        }
         $post->delete();
 
         return redirect('/posts')->with('success', 'Post deleted successfully!');
